controller and main refactor
