+++ src/utils/mod.rs
@@ -1,3 +1,5 @@
-pub mod server;
+#[cfg(feature = "server")]
 pub mod serverfn;
 pub mod table;
+#[cfg(feature = "server")]
+pub(crate) mod config;
